third_party/blink/web_tests/http/tests/websocket/send-file-blob-fail-expected.txt

WebSocket should fail the connection if it has failed to read a Blob.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Got FileSystem object.
File created.
PASS fileEntry.isFile is true
Wrote to file.
Got File object.
File deleted.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
